BERT model, translator, urdu labels and captions
This repository portrays our attempt at implementing the BERT architecture for obtaining urdu captions based on viewed images. The project was partially successful, with urdu labels being successfully portrayed but the captions were not as precise as assumed. A few bugs in the code, but can be tweaked for finalized results.